+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `duration` to scores
 - Restructure `personalBest` property to expose `outcome` directly in the `POST /login` response
 - Update integration tests
+- `PUT /player`: Change `dateOfBirth` into optional
 
 ## 0.2.1 - 02/02/2026 13:30 UTC+1